feat(movies): correções na atualização de filmes na base de dados

+++ src/main/java/cs2024/inf/CineView/dto/movies/MovieDto.java
@@ -20,7 +20,7 @@ public class MovieDto {
     private String tagline;
     private Double popularity;
     private Double vote_average;
-    private Double vote_count;
+    private Long vote_count;
     private List<GenreDto> genreModels;
 
 }
